Refactor AppLauncher into hooks and components
